API/flask_api/requirements.txt
psycopg2-binary==2.9.7
pydantic==2.3.0
python-dateutil==2.8.2
flask[async]==2.3.2
asyncio==3.4.3
pandas==2.0.3
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.1